+++ src/app/plugins/account/components/theme/actions.ts
@@ -1,4 +1,7 @@
 import ApiClient from 'api-client/ApiClient'
+import { preferencesActions } from 'core/session/preferencesReducers'
+import { themeActions } from 'core/session/themeReducers'
+import { AppPlugins } from 'app/constants'
 
 const { preferenceStore } = ApiClient.getInstance()
 
@@ -13,3 +16,35 @@ export const updateThemeConfig = async (config) =>
   preferenceStore.updateGlobalPreference('theme', config)
 
 export const deleteThemeConfig = async () => preferenceStore.deleteGlobalPreference('theme')
+
+export const updateSessionTheme = () => {
+  return (dispatch, settings, theme) => {
+    const themeUpdateBody = {
+      components: [
+        { pathTo: ['header', 'background'], value: settings.headerHex || theme.palette.grey[900] },
+        {
+          pathTo: ['sidebar', AppPlugins.MyAccount, 'background'],
+          value: settings.sidenavHex || theme.palette.grey[200],
+        },
+        {
+          pathTo: ['sidebar', AppPlugins.Kubernetes, 'background'],
+          value: settings.sidenavHex || theme.palette.grey[800],
+        },
+        {
+          pathTo: ['sidebar', AppPlugins.OpenStack, 'background'],
+          value: settings.sidenavHex || theme.palette.grey[800],
+        },
+        {
+          pathTo: ['sidebar', AppPlugins.BareMetal, 'background'],
+          value: settings.sidenavHex || theme.palette.grey[800],
+        },
+      ],
+    }
+    dispatch(
+      preferencesActions.updateLogo({
+        logoUrl: settings.logoUrl,
+      }),
+    )
+    dispatch(themeActions.updateThemeComponent(themeUpdateBody))
+  }
+}
